requirements.txt
# Core
google-genai>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Reliability
tenacity>=8.2.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
